add Scala 3 support
