+++ src/it/scala/zio/cassandra/session/SessionSpec.scala
@@ -103,9 +103,8 @@ object SessionSpec extends ZIOCassandraSpec with ZIOCassandraSpecUtils {
     },
     test("selectFirst should return Some(null) for null") {
       for {
-        result <- Session
-                    .selectFirst(s"select data FROM $keyspace.test_data WHERE id = 0")
-                    .map(_.map(_.getString(0)))
+        session <- ZIO.service[Session]
+        result <- session.selectFirst(s"select data FROM $keyspace.test_data WHERE id = 0").map(_.map(_.getString(0)))
       } yield assertTrue(result.contains(null))
     },
     test("selectFirst interpolated query (cqlConst) should return Some") {
